refactor: decompose call_bloc.dart into focused part files
Extract three groups of methods from the monolithic call_bloc.dart into
dedicated part files, reducing the root file from ~1050 to ~340 lines:

- call_observers.dart: onChange side-effect helpers (_syncCallkeepSignalingStatus,
  _handleBackgroundConnectivity, _syncPeerConnections, _logProcessingStatusTransitions,
  _handleRegistrationChange, _syncLinesState, _handleCallEndedCallback,
  _handleCallLifecycleTransitions, _handleSignalingSessionError, _notifyAccountErrorSafely)

- call_lifecycle.dart: app/connectivity/push/reset event handlers
  (_onCallStarted, _onAppLifecycleStateChanged, _onConnectivityResultChanged,
  _onNavigatorMediaDevicesChange, _onRegistrationChange, _onResetStateEvent,
  _onCallPushEventIncoming, _onConfigEvent)

- call_control.dart: CallControlEvent and CallScreenEvent handlers
  (all __onCallControl* and __onCallScreenEvent* methods)

call_bloc.dart now reads as a declaration: fields, constructor with event
registrations, close(), onChange() dispatch calls, and delegate methods.
