+++ lib/features/call/bloc/call_lifecycle.dart
@@ -1,0 +1,205 @@
+part of 'call_bloc.dart';
+
+extension _CallLifecycle on CallBloc {
+  Future<void> _onCallStarted(CallStarted event, Emitter<CallState> emit) async {
+    AppleNativeAudioManagement.setUseManualAudio(true);
+
+    // Initialize app lifecycle state
+    final lifecycleState = WidgetsFlutterBinding.ensureInitialized().lifecycleState;
+    emit(state.copyWith(currentAppLifecycleState: lifecycleState));
+    _logger.fine('_onCallStarted initial lifecycle state: $lifecycleState');
+
+    // Initialize connectivity state
+    final connectivityState = (await Connectivity().checkConnectivity()).first;
+    emit(
+      state.copyWith(
+        callServiceState: state.callServiceState.copyWith(networkStatus: connectivityState.toNetworkStatus()),
+      ),
+    );
+    _logger.finer('_onCallStarted initial connectivity state: $connectivityState');
+
+    // Subscribe to future connectivity changes
+    _connectivityChangedSubscription = Connectivity().onConnectivityChanged.listen((result) {
+      final currentConnectivityResult = result.first;
+      add(_ConnectivityResultChanged(currentConnectivityResult));
+    });
+
+    _signalingModule.reconnect(delay: Duration.zero);
+
+    WebRTC.initialize(options: webRtcOptionsBuilder?.build());
+  }
+
+  Future<void> _onAppLifecycleStateChanged(_AppLifecycleStateChanged event, Emitter<CallState> emit) async {
+    final appLifecycleState = event.state;
+    _logger.fine('_onAppLifecycleStateChanged: $appLifecycleState');
+
+    emit(state.copyWith(currentAppLifecycleState: appLifecycleState));
+
+    if (appLifecycleState == AppLifecycleState.paused || appLifecycleState == AppLifecycleState.detached) {
+      if (state.isActive == false) _signalingModule.disconnect();
+    } else if (appLifecycleState == AppLifecycleState.resumed) {
+      _signalingModule.reconnect();
+    }
+  }
+
+  Future<void> _onConnectivityResultChanged(_ConnectivityResultChanged event, Emitter<CallState> emit) async {
+    final connectivityResult = event.result;
+    _logger.fine('_onConnectivityResultChanged: $connectivityResult');
+    if (connectivityResult == ConnectivityResult.none) {
+      _signalingModule.disconnect();
+    } else {
+      _signalingModule.reconnect();
+    }
+    emit(
+      state.copyWith(
+        callServiceState: state.callServiceState.copyWith(networkStatus: connectivityResult.toNetworkStatus()),
+      ),
+    );
+  }
+
+  Future<void> _onNavigatorMediaDevicesChange(_NavigatorMediaDevicesChange event, Emitter<CallState> emit) async {
+    if (Platform.isIOS) {
+      // Cleanup devices info if change happened after hangup
+      // to avoid presenting stale data on next call initialization
+      if (state.activeCalls.isEmpty) return emit(state.copyWith(availableAudioDevices: [], audioDevice: null));
+
+      final devices = await navigator.mediaDevices.enumerateDevices();
+      final output = devices.where((d) => d.kind == 'audiooutput').toList();
+      final input = devices.where((d) => d.kind == 'audioinput').toList();
+      _logger.info('Devices change - out:${output.map((e) => e.str).toList()}, in:${input.map((e) => e.str).toList()}');
+
+      final available = [
+        CallAudioDevice(type: CallAudioDeviceType.speaker),
+        ...input.map(CallAudioDevice.fromMediaInput),
+      ];
+
+      CallAudioDevice current;
+
+      if (output.isNotEmpty) {
+        current = CallAudioDevice.fromMediaOutput(output.first);
+      } else {
+        // Fallback behavior for iOS when out:[]
+        // We prioritize the Earpiece (Receiver) if available (derived from MicrophoneBuiltIn),
+        // otherwise fallback to the first available device (which is Speaker based on the list above).
+        current = available.firstWhere(
+          (device) => device.type == CallAudioDeviceType.earpiece,
+          orElse: () => available.first,
+        );
+
+        _logger.warning(
+          'No "audiooutput" devices reported. Fallback selected: ${current.name} (type: ${current.type})',
+        );
+      }
+
+      emit(state.copyWith(availableAudioDevices: available, audioDevice: current));
+    }
+  }
+
+  // processing the registration event change
+
+  Future<void> _onRegistrationChange(_RegistrationChange event, Emitter<CallState> emit) async {
+    emit(state.copyWith(callServiceState: state.callServiceState.copyWith(registration: event.registration)));
+  }
+
+  // processing the handling of the app state
+  Future<void> _onResetStateEvent(_ResetStateEvent event, Emitter<CallState> emit) {
+    return switch (event) {
+      _ResetStateEventCompleteCalls() => __onResetStateEventCompleteCalls(event, emit),
+      _ResetStateEventCompleteCall() => __onResetStateEventCompleteCall(event, emit),
+    };
+  }
+
+  Future<void> __onResetStateEventCompleteCalls(_ResetStateEventCompleteCalls event, Emitter<CallState> emit) async {
+    _logger.warning('__onResetStateEventCompleteCalls: ${state.activeCalls}');
+
+    for (var element in state.activeCalls) {
+      // Skip outgoing calls not yet accepted — they are waiting for performStartCall
+      // to run, which will check registration and emit the appropriate notification.
+      // This mirrors the guard in _processHandshakeAsync that preserves the same calls.
+      if (element.direction == CallDirection.outgoing && element.acceptedTime == null && element.hungUpTime == null) {
+        continue;
+      }
+      add(_ResetStateEvent.completeCall(element.callId));
+    }
+  }
+
+  Future<void> __onResetStateEventCompleteCall(_ResetStateEventCompleteCall event, Emitter<CallState> emit) async {
+    _logger.warning('__onResetStateEventCompleteCall: ${event.callId}');
+
+    try {
+      emit(
+        state.copyWithMappedActiveCall(event.callId, (activeCall) {
+          return activeCall.copyWith(processingStatus: CallProcessingStatus.disconnecting);
+        }),
+      );
+
+      await state.performOnActiveCall(event.callId, (activeCall) async {
+        // Retrieve PC via manager and close it
+        await _peerConnectionManager.disposePeerConnection(activeCall.callId);
+
+        await callkeep.reportEndCall(
+          activeCall.callId,
+          activeCall.displayName ?? activeCall.handle.value,
+          CallkeepEndCallReason.remoteEnded,
+        );
+        await activeCall.localStream?.dispose();
+      });
+      emit(state.copyWithPopActiveCall(event.callId));
+    } on Error {
+      rethrow;
+    } catch (e, s) {
+      _logger.warning('__onResetStateEventCompleteCall', e, s);
+    }
+  }
+
+  // processing call push events
+
+  Future<void> _onCallPushEventIncoming(_CallPushEventIncoming event, Emitter<CallState> emit) async {
+    final eventError = event.error;
+    if (eventError != null) {
+      _logger.warning('_onCallPushEventIncoming event.error: $eventError');
+      // TODO: implement correct incoming call hangup (take into account that _signalingClient is disconnected)
+      return;
+    }
+
+    final contactName = await contactNameResolver.resolveWithNumber(event.handle.value);
+    final displayName = contactName ?? event.displayName;
+
+    emit(
+      state.copyWithPushActiveCall(
+        ActiveCall(
+          direction: CallDirection.incoming,
+          line: _kUndefinedLine,
+          callId: event.callId,
+          handle: event.handle,
+          displayName: displayName,
+          video: event.video,
+          createdTime: clock.now(),
+          processingStatus: CallProcessingStatus.incomingFromPush,
+        ),
+      ),
+    );
+
+    // Replace the display name in Callkeep if it differs from the one in the event
+    // mostly needed for ios, coz android can do it on background fcm isolate directly before push
+    // TODO:
+    // - do it on backend side same as for messaging
+    //   currently push notification contain display name from sip header
+    if (displayName != event.displayName) {
+      await callkeep.reportUpdateCall(event.callId, displayName: displayName);
+    }
+
+    // Function to verify speaker availability for the upcoming event, ensuring the speaker button is correctly enabled or disabled
+    add(const _NavigatorMediaDevicesChange());
+
+    // the rest logic implemented within _onSignalingStateHandshake on IncomingCallEvent from call logs processing
+  }
+
+  void _onConfigEvent(CallConfigEvent event, Emitter<CallState> emit) {
+    switch (event) {
+      case _CallConfigEventUpdated(monitorCheckInterval: final interval):
+        _logger.info('Updating PeerConnectionManager configuration: monitorCheckInterval=$interval');
+        _peerConnectionManager.updateConfig(monitorCheckInterval: interval);
+    }
+  }
+}
